ipc_channel

Module ipc

Structs

IpcBytesReceiver

Receiving end of a channel that does not used serialized messages.

IpcBytesSender

Sending end of a channel that does not used serialized messages.

IpcMessage

Structure used to represent a raw message from an IpcSender.

IpcOneShotServer

A server associated with a given name. The server is “one-shot” because it accepts only one connect request from a client.

IpcReceiver

Receiving end of a channel using serialized messages.

IpcReceiverSet

Collection of IpcReceivers moved into the set; thus creating a common (and exclusive) endpoint for receiving messages on any of the added channels.

IpcSender

Sending end of a channel using serialized messages.

IpcSharedMemory

Shared memory descriptor that will be made accessible to the receiver of an IPC message that contains the discriptor.

OpaqueIpcReceiver

OpaqueIpcSender

Enums

IpcSelectionResult

Result for readable events returned from IpcReceiverSet::select.

Constants

OS_IPC_CHANNELS_FOR_DESERIALIZATION 🔒

OS_IPC_CHANNELS_FOR_SERIALIZATION 🔒

OS_IPC_SHARED_MEMORY_REGIONS_FOR_DESERIALIZATION 🔒

OS_IPC_SHARED_MEMORY_REGIONS_FOR_SERIALIZATION 🔒

Functions

bytes_channel

Create a connected IpcBytesSender and IpcBytesReceiver.

channel

Create a connected IpcSender and IpcReceiver that transfer messages of a given type provided by type T or inferred by the types of messages sent by the sender.

deserialize_os_ipc_receiver 🔒

deserialize_os_ipc_sender 🔒

serialize_os_ipc_receiver 🔒

serialize_os_ipc_sender 🔒