ipc_channel::platform::unix

Struct OsIpcSender

pub struct OsIpcSender {
    fd: Arc<SharedFileDescriptor>,
}

Fields
fd: Arc<SharedFileDescriptor>
Implementations

impl OsIpcSender

fn from_fd(fd: c_int) -> OsIpcSender

fn get_system_sendbuf_size(&self) -> Result<usize, UnixError>

Maximum size of the kernel buffer used for transfers over this channel.

Note: This is not the actual maximal packet size we are allowed to use… Some of it is reserved by the kernel for bookkeeping.

fn fragment_size(sendbuf_size: usize) -> usize

Calculate maximum payload data size per fragment.

It is the total size of the kernel buffer, minus the part reserved by the kernel.

The sendbuf_size passed in should usually be the maximum kernel buffer size, i.e. the value of *SYSTEM_SENDBUF_SIZE – except after getting ENOBUFS, in which case it needs to be reduced.

fn first_fragment_size(sendbuf_size: usize) -> usize

Calculate maximum payload data size of first fragment.

This one is smaller than regular fragments, because it carries the message (size) header.

pub fn get_max_fragment_size() -> usize

Maximum data size that can be transferred over this channel in a single packet.

This is the size of the main data chunk only – it’s independent of any auxiliary data (FDs) transferred along with it.

A send on this channel won’t block for transfers up to this size under normal circumstances. (It might still block if heavy memory pressure causes ENOBUFS, forcing us to reduce the packet size.)

pub fn send( &self, data: &[u8], channels: Vec<OsIpcChannel>, shared_memory_regions: Vec<OsIpcSharedMemory>, ) -> Result<(), UnixError>

pub fn connect(name: String) -> Result<OsIpcSender, UnixError>