Module ipc_channel::refcell

Structs

BorrowRef

BorrowRefMut

Ref
Wraps a borrowed reference to a value in a RefCell box. A wrapper type for an immutably borrowed value from a RefCell<T>.

RefCell
A fork of std::cell::RefCell that makes borrow_state usable on stable Rust.

RefMut
A wrapper type for a mutably borrowed value from a RefCell<T>.

Enums

BorrowState
An enumeration of values returned from the state method on a RefCell<T>.

Constants

UNUSED

WRITING

Type Definitions

BorrowFlag