Struct ipc_channel::refcell::RefCell

pub struct RefCell<T: ?Sized> {
    borrow: Cell<usize>,
    value: UnsafeCell<T>,
}

A fork of std::cell::RefCell that makes borrow_state usable on stable Rust.

FIXME(https://github.com/rust-lang/rust/issues/27708): Remove this (revert commit f7f81e0ed0b62402db291e28a9bb16f7194ebf78 / PR #11835) when std::cell::RefCell::as_unsafe_cell is in Rust’s stable channel.

Fields
borrow: Cell<usize> value: UnsafeCell<T>
Methods

impl<T> RefCell<T>

fn new(value: T) -> RefCell<T>

Creates a new RefCell containing value.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

impl<T: ?Sized> RefCell<T>

fn borrow_state(&self) -> BorrowState

Query the current state of this RefCell

The returned value can be dispatched on to determine if a call to borrow or borrow_mut would succeed.

fn borrow(&self) -> Ref<T>

Immutably borrows the wrapped value.

The borrow lasts until the returned Ref exits scope. Multiple immutable borrows can be taken out at the same time.

Panics

Panics if the value is currently mutably borrowed.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

let borrowed_five = c.borrow();
let borrowed_five2 = c.borrow();

An example of panic:

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow_mut();

   let b = c.borrow(); // this causes a panic
}).join();

assert!(result.is_err());

fn borrow_mut(&self) -> RefMut<T>

Mutably borrows the wrapped value.

The borrow lasts until the returned RefMut exits scope. The value cannot be borrowed while this borrow is active.

Panics

Panics if the value is currently borrowed.

Examples

use std::cell::RefCell;

let c = RefCell::new(5);

*c.borrow_mut() = 7;

assert_eq!(*c.borrow(), 7);

An example of panic:

use std::cell::RefCell;
use std::thread;

let result = thread::spawn(move || {
   let c = RefCell::new(5);
   let m = c.borrow();

   let b = c.borrow_mut(); // this causes a panic
}).join();

assert!(result.is_err());