ipc_channel

Module router

Routers allow converting IPC channels to crossbeam channels. The RouterProxy provides various methods to register IpcReceiver<T>s. The router will then either call the appropriate callback or route the message to a crossbeam Sender<T> or Receiver<T>. You should use the global ROUTER to access the RouterProxy methods (via ROUTER’s Deref for RouterProxy.

Structs

Router 🔒

Router runs in its own thread listening for events. Adds events to its IpcReceiverSet and listens for events using select().

RouterProxy

A RouterProxy provides methods for talking to the router. Calling new automatically spins up a router thread which waits for events on its registered IpcReceiver<T>s. The RouterProxy’s methods communicate with the running router thread to register new IpcReceiver<T>’s

RouterProxyComm 🔒

Enums

RouterHandler 🔒

RouterMsg 🔒

Statics

ROUTER

Global object wrapping a RouterProxy. Add routes (add_typed_route), or convert IpcReceiver<T> to crossbeam channels (e.g. route_ipc_receiver_to_new_crossbeam_receiver)

Type Aliases

RouterMultiHandler

Function to call when a new event is received from the corresponding receiver.

RouterOneShotHandler

Function to call the first time that a message is received from the corresponding receiver.

TypedRouterMultiHandler

Like RouterMultiHandler but includes the type that will be passed to the callback

TypedRouterOneShotHandler

Like RouterOneShotHandler but includes the type that will be passed to the callback