Module ipc_channel::ipc

Structs

IpcBytesReceiver
Receiving end of a channel that does not used serialized messages.

IpcBytesSender
Sending end of a channel that does not used serialized messages.

IpcOneShotServer
A server associated with a given name.

IpcReceiver
Receiving end of a channel using serialized messages.

IpcReceiverSet
Collection of IpcReceivers moved into the set; thus creating a common (and exclusive) endpoint for receiving messages on any of the added channels.

IpcSender
Sending end of a channel using serialized messages.

IpcSharedMemory
Shared memory descriptor that will be made accessible to the receiver of an IPC message that contains the discriptor.

OpaqueIpcMessage
Structure used to represent a raw message from an IpcSender.

OpaqueIpcReceiver

OpaqueIpcSender

Enums

IpcError

IpcSelectionResult
Result for readable events returned from IpcReceiverSet::select.

TryRecvError

Functions

bytes_channel
Create a connected IpcBytesSender and IpcBytesReceiver.

channel
Create a connected IpcSender and IpcReceiver that transfer messages of a given type privided by type T or inferred by the types of messages sent by the sender.